+++ src/main/java/com/clearsys/professorama/api/security/JwtUser.java
@@ -9,19 +9,19 @@ public class JwtUser implements UserDetails{
 	
 	private static final long serialVersionUID = -268046329085485932L;
 
-	private Long id;
+	private int id;
 	private String username;
 	private String password;
 	private Collection<? extends GrantedAuthority> authorities;
 	
-	public JwtUser(Long id, String username, String password, Collection<? extends GrantedAuthority> authorities) {
+	public JwtUser(int id, String username, String password, Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.username = username;
 		this.password = password;
 		this.authorities = authorities;
 	}
 	
-	public Long getId() {
+	public int getId() {
 		return this.id;
 	}
